refactor: rename IT modules to observability-kit-tests-* and extract shared test base
- Aggregator artifactId: observability-kit-micrometer-test → observability-kit-tests
- New observability-kit-test-common (jar) with AbstractIT in pkg
  com.vaadin.observability.tests.common; compile-scope testbench deps
- Standalone IT dir renamed to observability-kit-tests-micrometer,
  artifactId observability-kit-tests-micrometer, pkg
  com.vaadin.observability.tests.micrometer; depends on test-common
