+++ observability-kit-tests/observability-kit-tests-micrometer/src/main/frontend/index.html
