+++ observability-kit-tests/pom.xml
@@ -6,13 +6,14 @@
     <artifactId>observability-kit</artifactId>
     <version>5.0-SNAPSHOT</version>
   </parent>
-  <artifactId>observability-kit-micrometer-test</artifactId>
+  <artifactId>observability-kit-tests</artifactId>
 
   <packaging>pom</packaging>
-  <name>Integration tests for Observability-kit Micrometer integration</name>
+  <name>Integration tests for Observability-kit</name>
 
   <modules>
-    <module>observability-kit-micrometer-tests</module>
+    <module>observability-kit-test-common</module>
+    <module>observability-kit-tests-micrometer</module>
   </modules>
 
   <dependencyManagement>
